- add common dependencies and config. - structure refactoring
