+++ app/src/main/java/org/smartregister/sample/fp/app/FPApplication.kt
@@ -1,0 +1,9 @@
+package org.smartregister.sample.fp.app
+
+import org.smartregister.view.activity.DrishtiApplication
+
+class FPApplication : DrishtiApplication() {
+    override fun logoutCurrentUser() {
+
+    }
+}
