+++ familyplanningmodule/src/main/AndroidManifest.xml
@@ -2,11 +2,14 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="org.smartregister.fp">
 
-    <application>
-        <activity android:name=".login.ui.LoginActivity"></activity>
-        <meta-data
-            android:name="preloaded_fonts"
-            android:resource="@array/preloaded_fonts" />
+    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+
+    <application
+        android:clearTaskOnLaunch="true"
+        android:label="@string/app_name"
+        android:largeHeap="true"
+        android:supportsRtl="true">
     </application>
 
 </manifest>
