+++ familyplanningmodule/src/main/java/org/smartregister/fp/common/util/DBConstantsUtil.kt
@@ -1,0 +1,13 @@
+package org.smartregister.fp.common.util
+
+object DBConstantsUtil {
+
+    object OrderBy {
+        const val ASC = "ASC"
+        const val DESC = "DESC"
+    }
+
+    object ColumnInfo {
+        const val REGISTRATION_DATA = "registration_date"
+    }
+}
