+++ settings.gradle
@@ -1,3 +1,2 @@
-rootProject.name='Family Planning'
 include ':app'
 include ':familyplanningmodule'
